+++ playground/app/components/content/NuxtImg.ts
@@ -1,6 +1,6 @@
 import { h } from 'vue'
 import NuxtImg from '../../../node_modules/@nuxt/image/dist/runtime/components/nuxt-img.mjs'
 
-export default function render (attrs: Record<string, any>) {
+export default function render (attrs: Record<string, unknown>) {
   return h(NuxtImg, attrs)
 }