+++ playground/modules/list/module.ts
@@ -1,11 +1,11 @@
-import { join } from 'path'
+import { join } from 'node:path'
 import { defineNuxtModule } from '@nuxt/kit'
 import type { Nuxt } from '@nuxt/schema'
 
 console.log('list module imported')
 
 export default defineNuxtModule({
-  setup (_options: any, nuxt: Nuxt) {
+  setup (_options: unknown, nuxt: Nuxt) {
     nuxt.options.nitro.externals = nuxt.options.nitro.externals || {}
     nuxt.options.nitro.externals.inline = nuxt.options.nitro.externals.inline || []
     nuxt.options.nitro.externals.inline.push(join(__dirname, 'module.ts'))