+++ src/runtime/utils/debug.ts
@@ -8,6 +8,6 @@ export function warn (...data: any[]): void {
   console.warn(label, ...data)
 }
 
-export function list(message: string, items: string[]) {
+export function list (message: string, items: string[]) {
   log(`${message}:\n\n${items.map(item => `   - ${item}`).join('\n')}\n`)
 }